# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Damga API
  description: |
    Şeffaf İşyeri Yoklama platformu — public HTTP API.

    Tam entegrasyon rehberi: [api-readme.md](./api-readme.md)
    Webhook imza doğrulama: [webhook-verify.md](./webhook-verify.md)
  version: '1.0.0'
  contact:
    name: Damga Destek
    email: destek@damga.deploi.net
  license:
    name: Proprietary
servers:
  - url: https://api.damga.deploi.net/v1
    description: Production

# ─── Güvenlik tipleri ──────────────────────────────────────────────────
components:
  securitySchemes:
    SupabaseJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Supabase Auth oturum token'ı. Web/mobil client'lar için.
        Format: `Authorization: Bearer eyJhbGciOiJIUzI1NiIs...`

    OrgAdminKey:
      type: http
      scheme: bearer
      description: |
        Org owner tarafından `/v1/api-keys` ile üretilen API key.
        Format: `Authorization: Bearer dmg_live_...` veya `X-API-Key: dmg_live_...`
        Kendi org'una bağlı, scope ile sınırlı.

    ServiceKey:
      type: http
      scheme: bearer
      description: |
        Platform admin tarafından üretilen S2S key (Lokma vb. iç projeler).
        Format: `Authorization: Bearer dmg_svc_...`
        **Her istekte `?org_id=<uuid>` query param veya `X-Damga-Org` header ZORUNLU.**
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: 'İnsan-okunabilir mesaj (Türkçe)'
        code:
          type: string
          description: 'Machine-readable hata kodu (örn. RATE_LIMIT_EXCEEDED)'
        details:
          type: object
          additionalProperties: true
    Org:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        plan:
          type: string
          enum: [free, starter, pro, business, enterprise]
...
    ApiKey:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        key_prefix: { type: string, description: 'Görünen prefix (örn. dmg_live_abc123...)' }
        scopes:
          type: array
          items: { type: string }
        rate_limit_per_min: { type: integer }
        last_used_at: { type: string, format: date-time, nullable: true }
        is_active: { type: boolean }
    ServiceKeyCreateRequest:
      type: object
      required: [name, scopes]
      properties:
        name: { type: string, minLength: 2, maxLength: 100 }
        scopes:
          type: array
          minItems: 1
          items: { type: string }
        rate_limit_per_min:
          type: integer
          minimum: 1
          maximum: 10000
        expires_at:
          type: string
          format: date-time
          nullable: true
    ServiceKeyCreateResponse:
      type: object
      properties:
        service_key: { $ref: '#/components/schemas/ApiKey' }
        secret_key:
          type: string
          description: 'Raw key (`dmg_svc_...`) — sadece bu yanıtta döner. Bir daha gösterilmez.'
        warning: { type: string }
  responses:
...
    Forbidden:
      description: 'Yetki yok / scope eksik'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /platform/service-keys:
    get:
      summary: 'Service key listesi (sadece platform admin)'
      tags: [Platform]
      security:
        - SupabaseJWT: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ApiKey' }

    post:
      summary: 'Yeni service key üret (sadece platform admin)'
      tags: [Platform]
      security:
        - SupabaseJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ServiceKeyCreateRequest' }
      responses:
        '201':
          description: 'Raw key sadece burada döner'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServiceKeyCreateResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /platform/orgs:
    get:
      summary: 'Tüm org listesi (sadece platform admin)'
      tags: [Platform]
      security:
        - SupabaseJWT: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Org'
                        - type: object
                          properties:
                            user_count: { type: integer }
                            location_count: { type: integer }
                            check_in_count: { type: integer }
                            last_activity: { type: string, format: date-time, nullable: true }
...